api/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
django-filter==23.5
python-telegram-bot==13.7
